fix padding in table and relocate component
